JSRPC: Fix regression when serializing arguments fails.
Recent changes made it so that we send the CustomEvent to the WorkerInterface before the top-level RPC call has actually been constructed. But this means if we fail to serialize the parameters, the event is already running, and it will just hang open since the server side won't return until the client side drops all the capabilities.

To fix this, we apply a membrane on the client side that cancels the requset when all capabilities have been dropped.

This feels ugly. You might argue we should instead further complexify the signature of getClientForOneCall() so that it returns some sort of callback that can be invoked to say "ok, I successfully initiated a call", and then only send the CustomEvent when that's called. That would recreate the previous behavior. However, that approach is not fully correct either: If for some reason the call is sent but fails to make it all the way to the server, the same hang can happen.

The new solution also has the advantage that if the last capability is dropped from the client side (which is probably the common case), then the event can be cleaned up immediately without waiting for a round trip to the server for the server to notice.

There is a disadvantage, which is that most RPC sessions will appear to end in cancellation (e.g. in metrics). However, this is not that big a deal and we can find a fix for it later.